requirements.txt
selenium
pillow
numpy
python-docx